Simplify query param approach
