+++ src/components/Table/TasksTable.tsx
@@ -29,6 +29,7 @@ interface TasksTableProps {
   onExpand?: (expanded: boolean) => void;
   onClickTaskLink?: (taskId: string) => void;
   sorts?: SortOrder[];
+  controlled?: boolean;
 }
 export const TasksTable: React.FC<TasksTableProps> = ({
   tasks,
@@ -36,6 +37,7 @@ export const TasksTable: React.FC<TasksTableProps> = ({
   onExpand,
   onClickTaskLink,
   sorts,
+  controlled = false,
 }) => (
   <Table
     data-cy="tasks-table"
@@ -43,7 +45,7 @@ export const TasksTable: React.FC<TasksTableProps> = ({
     pagination={false}
     columns={
       sorts
-        ? getColumnDefsControlled(sorts, onClickTaskLink)
+        ? getColumnDefsSort(sorts, controlled, onClickTaskLink)
         : getColumnDefs(onClickTaskLink)
     }
     dataSource={tasks}
@@ -94,8 +96,9 @@ const getColumnDefs = (onClickTaskLink): ColumnProps<Task>[] => [
   },
 ];
 
-const getColumnDefsControlled = (
+const getColumnDefsSort = (
   sortOrder: SortOrder[],
+  controlled: boolean,
   onClickTaskLink: (taskId: string) => void
 ): ColumnProps<Task>[] => {
   const getSortDir = (
@@ -109,30 +112,40 @@ const getColumnDefsControlled = (
     }
     return undefined;
   };
-  const sortProps = [
+  const controlledSortProps = [
     {
       sorter: {
         multiple: 4,
       },
-      sortOrder: getSortDir(TaskSortCategory.Name, sortOrder),
     },
     {
       sorter: {
         multiple: 4,
       },
-      sortOrder: getSortDir(TaskSortCategory.Status, sortOrder),
     },
     {
       dataIndex: ["baseTask", "status"],
       sorter: {
         multiple: 4,
       },
-      sortOrder: getSortDir(TaskSortCategory.BaseStatus, sortOrder),
     },
     {
       sorter: {
         multiple: 4,
       },
+    },
+  ];
+  const sortProps = [
+    {
+      sortOrder: getSortDir(TaskSortCategory.Name, sortOrder),
+    },
+    {
+      sortOrder: getSortDir(TaskSortCategory.Status, sortOrder),
+    },
+    {
+      sortOrder: getSortDir(TaskSortCategory.BaseStatus, sortOrder),
+    },
+    {
       sortOrder: getSortDir(TaskSortCategory.Variant, sortOrder),
     },
   ];
@@ -140,6 +153,7 @@ const getColumnDefsControlled = (
   return getColumnDefs(onClickTaskLink).map((columnDef, i) => ({
     ...columnDef,
     ...sortProps[i],
+    ...(controlled && controlledSortProps[i]),
   }));
 };
