+++ src/pages/patch/patchTabs/Tasks.tsx
@@ -1,8 +1,11 @@
+import { useEffect } from "react";
 import { useQuery } from "@apollo/client";
 import styled from "@emotion/styled";
 import Button from "@leafygreen-ui/button";
 import { Skeleton } from "antd";
-import { useParams, useHistory } from "react-router-dom";
+import every from "lodash.every";
+import queryString from "query-string";
+import { useParams, useLocation, useHistory } from "react-router-dom";
 import { usePatchAnalytics } from "analytics";
 import { PageSizeSelector } from "components/PageSizeSelector";
 import { Pagination } from "components/Pagination";
@@ -14,11 +17,14 @@ import { useToastContext } from "context/toast";
 import { PatchTasksQuery, PatchTasksQueryVariables } from "gql/generated/types";
 import { GET_PATCH_TASKS } from "gql/queries";
 import { useNetworkStatus } from "hooks";
-import { useGetTaskQueryVariables } from "hooks/useGetTaskQueryVariables";
-import { useTaskSortQueryParams } from "hooks/useTaskSortQueryParams";
+import { useUpdateURLQueryParams } from "hooks/useUpdateURLQueryParams";
 import { PatchTasksTable } from "pages/patch/patchTabs/tasks/PatchTasksTable";
 import { TaskFilters } from "pages/patch/patchTabs/tasks/TaskFilters";
+import { PatchTasksQueryParams, TaskStatus } from "types/task";
+import { url } from "utils";
+import { parseSortString } from "./util";
 
+const { getPageFromSearch, getLimitFromSearch } = url;
 interface Props {
   taskCount: number;
 }
@@ -26,13 +32,23 @@ interface Props {
 export const Tasks: React.FC<Props> = ({ taskCount }) => {
   const { id: resourceId } = useParams<{ id: string }>();
 
+  const { search } = useLocation();
   const router = useHistory();
   const patchAnalytics = usePatchAnalytics();
   const dispatchToast = useToastContext();
 
-  const queryVariables = useGetTaskQueryVariables();
-  const { limit, page } = queryVariables;
-  const sorts = useTaskSortQueryParams();
+  const updateQueryParams = useUpdateURLQueryParams();
+  const queryVariables = getQueryVariables(search, resourceId);
+
+  const { sorts, limit, page } = queryVariables;
+
+  useEffect(() => {
+    if (sorts.length === 0) {
+      updateQueryParams({
+        sorts: "STATUS:ASC;BASE_STATUS:DESC",
+      });
+    }
+  }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   const { data, startPolling, stopPolling } = useQuery<
     PatchTasksQuery,
@@ -97,6 +113,68 @@ export const Tasks: React.FC<Props> = ({ taskCount }) => {
       )}
     </>
   );
+};
+
+const getString = (param: string | string[]): string =>
+  Array.isArray(param) ? param[0] : param;
+const getArray = (param: string | string[]): string[] =>
+  Array.isArray(param) ? param : [param];
+
+const statusesToIncludeInQuery = {
+  [TaskStatus.Dispatched]: true,
+  [TaskStatus.Failed]: true,
+  [TaskStatus.Inactive]: true,
+  [TaskStatus.SetupFailed]: true,
+  [TaskStatus.Started]: true,
+  [TaskStatus.StatusBlocked]: true,
+  [TaskStatus.Succeeded]: true,
+  [TaskStatus.SystemFailed]: true,
+  [TaskStatus.SystemTimedOut]: true,
+  [TaskStatus.SystemUnresponsive]: true,
+  [TaskStatus.TaskTimedOut]: true,
+  [TaskStatus.TestTimedOut]: true,
+  [TaskStatus.Undispatched]: true,
+  [TaskStatus.Unstarted]: true,
+  [TaskStatus.Aborted]: true,
+};
+
+const getStatuses = (rawStatuses: string[] | string): string[] => {
+  const statuses = getArray(rawStatuses).filter(
+    (status) => status in statusesToIncludeInQuery
+  );
+  if (
+    every(Object.keys(statusesToIncludeInQuery), (status) =>
+      statuses.includes(status)
+    )
+  ) {
+    // passing empty array for `All` value is also more performant for filtering on the backend as opposed to passing array of all statuses
+    return [];
+  }
+  return statuses;
+};
+
+const getQueryVariables = (
+  search: string,
+  resourceId: string
+): PatchTasksQueryVariables => {
+  const {
+    [PatchTasksQueryParams.Variant]: variant,
+    [PatchTasksQueryParams.TaskName]: taskName,
+    [PatchTasksQueryParams.Statuses]: rawStatuses,
+    [PatchTasksQueryParams.BaseStatuses]: rawBaseStatuses,
+    [PatchTasksQueryParams.Sorts]: sorts,
+  } = queryString.parse(search, { arrayFormat: "comma" });
+
+  return {
+    patchId: resourceId,
+    sorts: parseSortString(sorts),
+    variant: getString(variant),
+    taskName: getString(taskName),
+    statuses: getStatuses(rawStatuses),
+    baseStatuses: getStatuses(rawBaseStatuses),
+    page: getPageFromSearch(search),
+    limit: getLimitFromSearch(search),
+  };
 };
 
 const FlexContainer = styled.div`
