+++ src/pages/patch/patchTabs/tasks/PatchTasksTable.tsx
@@ -24,6 +24,7 @@ export const PatchTasksTable: React.FC<Props> = ({ patchTasks, sorts }) => {
 
   return (
     <TasksTable
+      controlled
       sorts={sorts}
       tableChangeHandler={tableChangeHandler}
       tasks={patchTasks?.tasks}
