+++ src/pages/patch/patchTabs/util.test.ts
@@ -23,4 +23,22 @@ test("toSortString", () => {
     order: "descend",
   };
   expect(toSortString(input)).toEqual("NAME:DESC");
+
+  const unsetSort: SorterResult<Task> = {
+    columnKey: TaskSortCategory.Status,
+    order: undefined,
+  };
+  expect(toSortString(unsetSort)).toEqual(undefined);
+
+  const multiSort: SorterResult<Task>[] = [
+    {
+      columnKey: TaskSortCategory.Status,
+      order: undefined,
+    },
+    {
+      columnKey: TaskSortCategory.BaseStatus,
+      order: "ascend",
+    },
+  ];
+  expect(toSortString(multiSort)).toEqual("BASE_STATUS:ASC");
 });
